Fixed remote configuration

+++ src/main/java/org/thingsboard/gateway/service/DefaultTenantManagerService.java
@@ -1,12 +1,12 @@
 /**
  * Copyright © 2017 The Thingsboard Authors
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,14 +17,12 @@ package org.thingsboard.gateway.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import org.thingsboard.gateway.extensions.ExtensionService;
 import org.thingsboard.gateway.extensions.http.HttpService;
 import org.thingsboard.gateway.service.conf.TbExtensionConfiguration;
 import org.thingsboard.gateway.service.conf.TbGatewayConfiguration;
 import org.thingsboard.gateway.service.conf.TbTenantConfiguration;
 import org.thingsboard.gateway.service.gateway.GatewayService;
-import org.thingsboard.gateway.service.gateway.MqttGatewayService;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -43,7 +41,7 @@ public abstract class DefaultTenantManagerService implements TenantManagerServic
     @Autowired
     private TbGatewayConfiguration configuration;
 
-    private Map<String, TenantServicesRegistry> gateways;
+    private Map<String, TenantServiceRegistry> gateways;
     private List<HttpService> httpServices;
     private Boolean isRemoteConfiguration;
 
@@ -60,14 +58,18 @@ public abstract class DefaultTenantManagerService implements TenantManagerServic
             if (isRemoteConfiguration) {
                 String label = configuration.getLabel();
                 log.info("[{}] Initializing gateway", configuration.getLabel());
-                GatewayService service = getGatewayService(configuration, c -> onExtensionConfigurationUpdate(label, c));//new MqttGatewayService(configuration, c -> onExtensionConfigurationUpdate(label, c));
+                TenantServiceRegistry tenantServiceRegistry = new TenantServiceRegistry();
+                GatewayService service = null;
                 try {
-                    service.init();
-                    gateways.put(label, new TenantServicesRegistry(service));
+                    service = getGatewayService(configuration, c -> tenantServiceRegistry.updateExtensionConfiguration(c));
+                    tenantServiceRegistry.setService(service);
+                    gateways.put(label, tenantServiceRegistry);
                 } catch (Exception e) {
                     log.info("[{}] Failed to initialize the service ", label, e);
                     try {
-                        service.destroy();
+                        if (service != null) {
+                            service.destroy();
+                        }
                     } catch (Exception exc) {
                         log.info("[{}] Failed to stop the service ", label, exc);
                     }
@@ -75,23 +77,26 @@ public abstract class DefaultTenantManagerService implements TenantManagerServic
             } else {
                 String label = configuration.getLabel();
                 log.info("[{}] Initializing gateway", configuration.getLabel());
-                GatewayService service = getGatewayService(configuration, c -> {});//new MqttGatewayService(configuration, c -> {});
+                GatewayService service = null;
                 try {
-                    //service.init();
-                    ExtensionServiceCreation serviceCreation = new TenantServicesRegistry(service);
+                    TenantServiceRegistry tenantServiceRegistry = new TenantServiceRegistry();
+                    service = getGatewayService(configuration, c -> {});
+                    tenantServiceRegistry.setService(service);
                     for (TbExtensionConfiguration extensionConfiguration : configuration.getExtensions()) {
                         log.info("[{}] Initializing extension: [{}]", configuration.getLabel(), extensionConfiguration.getType());
-                        ExtensionService extension = serviceCreation.createExtensionServiceByType(service, extensionConfiguration.getType());
+                        ExtensionService extension = tenantServiceRegistry.createExtensionServiceByType(service, extensionConfiguration.getType());
                         extension.init(extensionConfiguration, isRemoteConfiguration);
                         if (extensionConfiguration.getType().equals("HTTP")) {
                             httpServices.add((HttpService) extension);
                         }
                     }
-                    gateways.put(label, (TenantServicesRegistry) serviceCreation);
+                    gateways.put(label, (TenantServiceRegistry) tenantServiceRegistry);
                 } catch (Exception e) {
                     log.info("[{}] Failed to initialize the service ", label, e);
                     try {
-                        service.destroy();
+                        if (service != null) {
+                            service.destroy();
+                        }
                     } catch (Exception exc) {
                         log.info("[{}] Failed to stop the service ", label, exc);
                     }
@@ -100,16 +105,10 @@ public abstract class DefaultTenantManagerService implements TenantManagerServic
         }
     }
 
-    private void onExtensionConfigurationUpdate(String label, String configuration) {
-        TenantServicesRegistry registry = gateways.get(label);
-        log.info("[{}] Updating extension configuration", label);
-        registry.updateExtensionConfiguration(configuration);
-    }
-
     @Override
     public void processRequest(String converterId, String token, String body) throws Exception {
         if (isRemoteConfiguration) {
-            for (TenantServicesRegistry tenant : gateways.values()) {
+            for (TenantServiceRegistry tenant : gateways.values()) {
                 tenant.processRequest(converterId, token, body);
             }
         } else {
@@ -123,7 +122,7 @@ public abstract class DefaultTenantManagerService implements TenantManagerServic
     public void stop() {
         for (String label : gateways.keySet()) {
             try {
-                TenantServicesRegistry registry = gateways.get(label);
+                TenantServiceRegistry registry = gateways.get(label);
                 for (ExtensionService extension : registry.getExtensions().values()) {
                     try {
                         if (isRemoteConfiguration) {
